% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curereg.R
\name{curefraction}
\alias{curefraction}
\title{Predict cure fractions}
\usage{
curefraction(object, newData, unique = TRUE, ordered = TRUE, n = 6)
}
\arguments{
\item{object}{an object of curereg.}

\item{newData}{um novo conjunto de dados com estrutura semelhante a utilizada no ajuste
(opcional). Se newData nao e especificado e utilizado o conjunto de dados usado
 no ajuste.}

\item{unique}{se igual a TRUE (unique = TRUE por padrao) sao escolhidos nos dados as linhas
de perfis unicos, ou seja, sao escolhidos apenas os clientes que possuem caracteristicas
diferentes.}

\item{ordered}{se igual a TRUE (ordered = TRUE por padrao) os dados sao ordenados em relacao
a fracao de cura.}

\item{n}{(n = 6 por padrao), argumento passado as funcoes \code{\link{head}} e
\code{\link{tail}} que informam quantas linhas do inicio e do fim do conjunto de
dados devem ser mostradas na saida da funcao.}
}
\value{
Retorna um data.frame dos dados adicionado de uma coluna com as fracoes de cura
(coluna curefraction).
}
\description{
Predict cure fractions from fitted \code{curereg} model.
}
\details{
O argumento \code{newData} quando utilizado deve ser estrutura de acordo com a saida
\code{coef(fit, terms = "cure")}, ou seja, variaveis que sao fatores devem ser convertidas em
dummy. Alem disso, deve ser considerada a ordem em que aparecem os coeficientes associados a
cada uma das variaveis.
}
\author{
Rumenick Pereira da Silva \email{rumenickps@gmail.com}
}
\keyword{models}
\keyword{predict}

