% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curereg.R
\name{lines.curereg}
\alias{lines.curereg}
\title{Add fitted flexible survival with cure fraction curves to a plot}
\usage{
\method{lines}{curereg}(x, ...)
}
\arguments{
\item{x}{an object of \code{\link{curereg}}.}

\item{...}{Other options to be passed to \code{\link{plot.flexsurvreg}}  or \code{\link{plot.survfit}}.}
}
\description{
Add fitted survival (or hazard or cumulative hazard) curves from a curereg
model fit to an existing plot.
}
\author{
Rumenick Pereira da Silva \email{rumenickps@gmail.com}
}
\seealso{
\code{\link{curereg}}, \code{\link{flexsurvreg}}
}
\keyword{models}
\keyword{plot}

