% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curereg.R
\name{plot.curereg}
\alias{plot.curereg}
\title{Plots of fitted flexible survival models with cure fraction}
\usage{
\method{plot}{curereg}(x, ...)
}
\arguments{
\item{x}{an object of \code{\link{curereg}}.}

\item{...}{Other options to be passed to \code{\link{plot.flexsurvreg}}  or \code{\link{plot.survfit}}.}
}
\description{
Plot fitted survival, cumulative hazard or hazard from a parametric model
against nonparametric estimates to diagnose goodness-of-fit.
}
\author{
Rumenick Pereira da Silva \email{rumenickps@gmail.com}
}
\seealso{
\code{\link{curereg}}, \code{\link{flexsurvreg}}
}
\keyword{models}
\keyword{plot}

