% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{LRT}
\alias{LRT}
\title{Likelihood ratio test}
\usage{
LRT(fitg, fits)
}
\arguments{
\item{fitg}{an object that stores the results of
curereg fit of the model under the null hypothesis.}

\item{fits}{an object that stores the results of curereg fit
of the model under the alternative hypothesis.}
}
\value{
\item{LRS }{
the value of the likelihood ratio statistic.
}
\item{pvalue }{
the p value of test under null hypothesis chi-square distribution.
}
}
\description{
Computes the likelihood ratio test.
}
\details{
The objects fitg and fits are obtained
using the usual options passed to the curereg
function.
}
\examples{

data(e1684)
fitg <- curereg(Surv(FAILTIME, FAILCENS) ~ 1, cureformula = ~ 1,
                 data = e1684, timedist = "moeev", ncausedist = "bernoulli")
fits <- curereg(Surv(FAILTIME, FAILCENS) ~ 1, cureformula = ~ 1,
                 data = e1684, timedist = "moee", ncausedist = "bernoulli")
LRT(fitg, fits)

}
\author{
Rumenick Pereira da Silva \email{rumenickps@gmail.com}
}
\keyword{test}

